Move wrapNumber
