+++ packages/parcels/src/parcelData/wrapNumber.js
@@ -1,0 +1,5 @@
+// @flow
+export default (val: number, max: number): number => {
+    val = val % max;
+    return val < 0 ? val + max : val;
+};
