--- packages/parcels/src/util/wrapNumber.js
@@ -1,5 +1,0 @@
-// @flow
-export default function NumberWrap(val: number, max: number): number {
-    val = val % max;
-    return val < 0 ? val + max : val;
-}
